# server/ulmapi/dto/create_course_response.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from ulmapi.dto.base_model_ import Model
from ulmapi import util


class CreateCourseResponse(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, created_at=None, course_id=None, username=None):  # noqa: E501
        """CreateCourseResponse - a model defined in OpenAPI

        :param created_at: The created_at of this CreateCourseResponse.  # noqa: E501
        :type created_at: datetime
        :param course_id: The course_id of this CreateCourseResponse.  # noqa: E501
        :type course_id: str
        :param username: The username of this CreateCourseResponse.  # noqa: E501
        :type username: str
        """
        self.openapi_types = {
            'created_at': datetime,
            'course_id': str,
            'username': str
        }

        self.attribute_map = {
            'created_at': 'created_at',
            'course_id': 'course_id',
            'username': 'username'
        }

        self._created_at = created_at
        self._course_id = course_id
        self._username = username

    @classmethod
    def from_dict(cls, dikt) -> 'CreateCourseResponse':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CreateCourseResponse of this CreateCourseResponse.  # noqa: E501
        :rtype: CreateCourseResponse
        """
        return util.deserialize_model(dikt, cls)

    @property
    def created_at(self):
        """Gets the created_at of this CreateCourseResponse.


        :return: The created_at of this CreateCourseResponse.
        :rtype: datetime
        """
        return self._created_at

    @created_at.setter
    def created_at(self, created_at):
        """Sets the created_at of this CreateCourseResponse.


        :param created_at: The created_at of this CreateCourseResponse.
        :type created_at: datetime
        """
        if created_at is None:
            raise ValueError("Invalid value for `created_at`, must not be `None`")  # noqa: E501

        self._created_at = created_at

    @property
    def course_id(self):
        """Gets the course_id of this CreateCourseResponse.


        :return: The course_id of this CreateCourseResponse.
        :rtype: str
        """
        return self._course_id

    @course_id.setter
    def course_id(self, course_id):
        """Sets the course_id of this CreateCourseResponse.


        :param course_id: The course_id of this CreateCourseResponse.
        :type course_id: str
        """
        if course_id is None:
            raise ValueError("Invalid value for `course_id`, must not be `None`")  # noqa: E501

        self._course_id = course_id

    @property
    def username(self):
        """Gets the username of this CreateCourseResponse.


        :return: The username of this CreateCourseResponse.
        :rtype: str
        """
        return self._username

    @username.setter
    def username(self, username):
        """Sets the username of this CreateCourseResponse.


        :param username: The username of this CreateCourseResponse.
        :type username: str
        """
        if username is None:
            raise ValueError("Invalid value for `username`, must not be `None`")  # noqa: E501

        self._username = username