# server/ulmapi/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: This is an OpenAPI specification for the University Life Manager backend.
    Request/response objects can be specified in components. Endpoints can be specified
    in paths.
  title: University Life Manager API
  version: 1.0.0
servers:
- url: /api/v1
security:
- bearerAuth: []
paths:
  /cat:
    get:
      operationId: cat_get
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Cat'
                type: array
          description: OK
        "401":
          description: Credentials are missing or invalid
      summary: Get all cats for the user
      tags:
      - example
      x-openapi-router-controller: ulmapi.controllers.impl.example_controller
    post:
      operationId: cat_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cat'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
          description: Created
        "401":
          description: Credentials are missing or invalid
      summary: Create a new cat for the user
      tags:
      - example
      x-openapi-router-controller: ulmapi.controllers.impl.example_controller
  /cat/{catId}:
    get:
      operationId: cat_cat_id_get
      parameters:
      - description: Cat ID
        explode: false
        in: path
        name: catId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
          description: OK
        "401":
          description: Credentials are missing or invalid
        "404":
          description: Entity could not be found
      summary: Get cat by id for the user
      tags:
      - example
      x-openapi-router-controller: ulmapi.controllers.impl.example_controller
    put:
      operationId: cat_cat_id_put
      parameters:
      - description: Cat ID
        explode: false
        in: path
        name: catId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Cat'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cat'
          description: OK
        "401":
          description: Credentials are missing or invalid
        "404":
          description: Entity could not be found
      summary: Replace a cat for the user
      tags:
      - example
      x-openapi-router-controller: ulmapi.controllers.impl.example_controller
  /login:
    post:
      operationId: login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCredentials'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
          description: OK
        "401":
          description: Credentials are missing or invalid
      security: []
      summary: Exchange username and password for access token
      x-openapi-router-controller: ulmapi.controllers.impl.default_controller
  /signup:
    post:
      operationId: signup_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupInfo'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessToken'
          description: OK
        "400":
          description: A malformed request could not be processed, e.g. a field must
            be unique or non-empty but was not
      security: []
      summary: Creates a new user (signup)
      x-openapi-router-controller: ulmapi.controllers.impl.default_controller
  /users/{username}/course:
    get:
      operationId: users_username_course_get
      parameters:
      - description: Username
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/CourseInfo'
                type: array
          description: OK
        "400":
          description: A malformed request could not be processed, e.g. a field must
            be unique or non-empty but was not
      summary: Get all courses for the user
      tags:
      - Course
      x-openapi-router-controller: ulmapi.controllers.impl.course_controller
    post:
      operationId: users_username_course_post
      parameters:
      - description: Username
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInfo'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCourseResponse'
          description: Course created
        "400":
          description: A malformed request could not be processed, e.g. a field must
            be unique or non-empty but was not
      summary: Create a new course for the user
      tags:
      - Course
      x-openapi-router-controller: ulmapi.controllers.impl.course_controller
  /users/{username}/course/{courseID}:
    delete:
      operationId: users_username_course_course_iddelete
      parameters:
      - description: Username
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      - description: Course ID
        explode: false
        in: path
        name: courseID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteCourseResponse'
          description: OK
        "400":
          description: A malformed request could not be processed, e.g. a field must
            be unique or non-empty but was not
        "401":
          description: Credentials are missing or invalid
        "404":
          description: Entity could not be found
      summary: Delete a course for the user
      tags:
      - Course
      x-openapi-router-controller: ulmapi.controllers.impl.course_controller
    get:
      operationId: users_username_course_course_idget
      parameters:
      - description: Username
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      - description: Course ID
        explode: false
        in: path
        name: courseID
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseInfo'
          description: OK
        "400":
          description: A malformed request could not be processed, e.g. a field must
            be unique or non-empty but was not
        "401":
          description: Credentials are missing or invalid
        "404":
          description: Entity could not be found
      summary: Get course by id for the user
      tags:
      - Course
      x-openapi-router-controller: ulmapi.controllers.impl.course_controller
    put:
      operationId: users_username_course_course_idput
      parameters:
      - description: Username
        explode: false
        in: path
        name: username
        required: true
        schema:
          type: string
        style: simple
      - description: Course ID
        explode: false
        in: path
        name: courseID
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseInfo'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateCourseResponse'
          description: OK
        "400":
          description: A malformed request could not be processed, e.g. a field must
            be unique or non-empty but was not
        "401":
          description: Credentials are missing or invalid
        "404":
          description: Entity could not be found
      summary: Replace a course for the user
      tags:
      - Course
      x-openapi-router-controller: ulmapi.controllers.impl.course_controller
components:
  responses:
    AuthorizationError:
      description: Credentials are missing or invalid
    BadRequest:
      description: A malformed request could not be processed, e.g. a field must be
        unique or non-empty but was not
    NotFoundError:
      description: Entity could not be found
  schemas:
    CourseInfo:
      example:
        expected_difficulty: 1
        course_id: course_id
        course_name: course_name
        desired_grade: 60
      properties:
        course_id:
          minLength: 1
          type: string
        course_name:
          minLength: 1
          type: string
        expected_difficulty:
          format: int32
          maximum: 5
          minimum: 1
          type: integer
        desired_grade:
          format: int32
          maximum: 100
          minimum: 0
          type: integer
      required:
      - course_id
      - course_name
      - desired_grade
      - expected_difficulty
      type: object
    CreateCourseResponse:
      example:
        course_id: course_id
        created_at: 2000-01-23T04:56:07.000+00:00
        username: username
      properties:
        created_at:
          format: date-time
          type: string
        course_id:
          type: string
        username:
          type: string
      required:
      - course_id
      - created_at
      - username
      type: object
    UpdateCourseResponse:
      example:
        course_id: course_id
        updated_at: 2000-01-23T04:56:07.000+00:00
        username: username
      properties:
        updated_at:
          format: date-time
          type: string
        course_id:
          type: string
        username:
          type: string
      required:
      - course_id
      - updated_at
      - username
      type: object
    DeleteCourseResponse:
      example:
        deleted: true
      properties:
        deleted:
          type: boolean
      required:
      - deleted
      type: object
    UserCredentials:
      example:
        password: password
        username: username
      properties:
        username:
          type: string
        password:
          type: string
      required:
      - password
      - username
      type: object
    AccessToken:
      example:
        access_token: access_token
      properties:
        access_token:
          type: string
      required:
      - access_token
      type: object
    SignupInfo:
      example:
        password: password
        email: email
        username: username
      properties:
        email:
          minLength: 1
          type: string
        username:
          minLength: 1
          type: string
        password:
          minLength: 1
          type: string
      required:
      - email
      - password
      - username
      type: object
    Cat:
      example:
        birth_date: 2000-01-23T04:56:07.000+00:00
        cat_id: cat_id
        name: name
        weight: 8
      properties:
        cat_id:
          type: string
        name:
          type: string
        weight:
          maximum: 100
          minimum: 0
          type: integer
        birth_date:
          format: date-time
          type: string
      required:
      - birth_date
      - name
      - weight
      type: object
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
      x-bearerInfoFunc: ulmapi.controllers.impl.security_controller_.info_from_bearerAuth